third_party/WebKit/LayoutTests/editing/execCommand/paste-2.html

<!DOCTYPE html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script>
var tester = async_test('iframe');
function runTest() {
    if (window.internals)
        internals.settings.setEditingBehavior('win');

    var frame = frames[0];
    var selection = frame.getSelection();
    var doc = frame.document;

    selection.collapse(doc.body, 0);
    doc.execCommand('InsertText', false, 'foo bar baz');
    selection.modify('extend', 'backward', 'word');
    doc.execCommand('Copy');
    doc.execCommand('Delete');
    doc.execCommand('Paste');

    tester.step(function() { assert_equals(doc.body.innerHTML.replace(/&nbsp;/, ' ', 'g'), 'foo bar baz'); });
    tester.done();
}
</script>
<p>This tests copy/delete/paste inside an editable iframe.  You should see 'foo bar baz' below.</p>
<iframe srcdoc="<body contenteditable></body>" onload="runTest()"></iframe>
<div id="log"></div>
